requirements.txt
flask
beautifulsoup4
requests
pillow
lxml
